remove commented out check if bin exists

+++ remote_syslog2/recipes/install.rb
@@ -5,7 +5,7 @@ remote_file install['download_path'] do
   Chef::Log.info("Installing remote_syslog2 from #{install['download_path']}...")
   source install['download_file']
   mode '0644'
-#  not_if { ::File.exists?(bin_file) }
+  not_if { ::File.exists?(bin_file) }
 end
 
 bash 'extract remote_syslog2' do
@@ -16,9 +16,7 @@ bash 'extract remote_syslog2' do
     mv #{install['extracted_path']}/#{install['extracted_bin']} #{bin_file}
     rm -rf #{install['download_path']} #{install['extracted_path']}
   EOH
-
-Chef::Log.info("Extract remote_syslog2...")
-#  not_if { ::File.exists?(bin_file) }
+  not_if { ::File.exists?(bin_file) }
 end
 
 file bin_file do
